add files quota on send message with files

+++ chat/application/users/files_quota.py
@@ -1,7 +1,9 @@
+from decimal import Decimal
+
 from core.constants import USER_FILES_QUOTA_MB
 
 
-def is_available_user_quota_for_file(user_current_files_size: float, file_size: float) -> bool:
+def is_available_user_quota_for_file(user_current_files_size: Decimal, file_size: Decimal) -> bool:
     """Доступна ли квота пользователя для загрузки файла.
 
     :param int user_current_files_size: Текущий размер всех файлов пользователя.
